+++ src/angular/angular-wizard.ts
@@ -172,6 +172,13 @@ Apologies for the inconvenience!`,
       )} to get a video-like reproduction of errors during a user session?`,
       enabledHint: 'recommended, but increases bundle size',
     },
+    {
+      id: 'logs',
+      prompt: `Do you want to enable ${chalk.bold(
+        'Logs',
+      )} to send your application logs to Sentry?`,
+      enabledHint: 'recommended',
+    },
   ] as const);
 
   await traceStep(